Implemented first simple test, moved GammaBernoulli
model code is now at gammabernoulli.py

trace["mu"] currently breaks, can be below zero, but should not.
