+++ plotting/check_prior_dist.py
