+++ settings.py
@@ -37,7 +37,7 @@ modes = [2, 1, 1, 1]
 inference = "NUTS"
 
 seed = 0  # for deterministic randomisation
-subset = 1  # only use every subset datapoint for bayes estimation for speedup
+subset = 10  # only use every subset datapoint for bayes estimation for speedup
 # use the estimated variability in qm
 scale_variability = True
 # out of "watch+wfdei", "GSWP3", "GSWP3+ERA5"
@@ -70,7 +70,6 @@ chains = 2  # number of chains to calculate (min 2 to check for convergence)
 # submitted jobs will have ncores_per_job=1 always.
 ncores_per_job = 2
 progressbar = True  # print progress in output (.err file for mpi)
-
 
 #### settings for create_submit.py
 # number of parallel jobs through jobarray
